pyproject.toml
[project]
name = "starfield-sim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "esper>=3.4",
    "pygame-ce>=2.5.5",
]
